+++ src/api/hooks/useGetThemes.ts
@@ -1,23 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
 
-import type { ThemeData } from '@/types';
-
-import { fetchInstance } from '../instance';
-
-export type ThemeResponseData = {
-  themes: ThemeData[];
-};
-
-const getThemesPath = () => '/v1/themes';
-const themesQueryKey = [getThemesPath()];
-
-export const getThemes = async () => {
-  const response = await fetchInstance.get<ThemeResponseData>(getThemesPath());
-  return response.data;
-};
+import { getThemes } from '../apis/theme';
+import API from '../path';
 
 export const useGetThemes = () =>
   useQuery({
-    queryKey: themesQueryKey,
+    queryKey: [API.THEME.GET_THEME],
     queryFn: getThemes,
   });
