+++ src/api/hooks/useGetThemesProducts.ts
@@ -4,40 +4,12 @@ import {
   type UseInfiniteQueryResult,
 } from '@tanstack/react-query';
 
-import type { GoodsData } from '@/types';
+import type { ThemeProductsRequestParams, ThemesProductsResponseData } from '../apis/theme';
+import { getThemesProducts } from '../apis/theme';
 
-import { fetchInstance } from '../instance';
-
-type RequestParams = {
-  themeKey: string;
-  pageToken?: string;
-  maxResults?: number;
+type Params = Pick<ThemeProductsRequestParams, 'maxResults' | 'themeKey'> & {
+  initPageToken?: string;
 };
-
-type ThemesProductsResponseData = {
-  products: GoodsData[];
-  nextPageToken?: string;
-  pageInfo: {
-    totalResults: number;
-    resultsPerPage: number;
-  };
-};
-
-const getThemesProductsPath = ({ themeKey, pageToken, maxResults }: RequestParams) => {
-  const params = new URLSearchParams();
-  if (pageToken) params.append('pageToken', pageToken);
-  if (maxResults) params.append('maxResults', maxResults.toString());
-  return `/v1/themes/${themeKey}/products?${params.toString()}`;
-};
-
-export const getThemesProducts = async (params: RequestParams) => {
-  const response = await fetchInstance.get<ThemesProductsResponseData>(
-    getThemesProductsPath(params),
-  );
-  return response.data;
-};
-
-type Params = Pick<RequestParams, 'maxResults' | 'themeKey'> & { initPageToken?: string };
 export const useGetThemesProducts = ({
   themeKey,
   maxResults = 20,
